% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metcolor.R
\name{metcolor}
\alias{metcolor}
\title{Color a metabolic map.}
\usage{
metcolor(svg, data, path = ".", quiet = FALSE, ...)
}
\arguments{
\item{svg}{SVG image file of the metabolic network, generated by MetDraw.}

\item{data}{Data frame with row names matching reaction names in the metabolic model.
One colored map will be produced for each column of data, using the column names to create
file names.}

\item{path}{Location for the output files.}

\item{quiet}{If \code{FALSE}, print updates as each file is processed.}

\item{colors}{Color pallette (list of color names).}

\item{normalize}{Normalization applied before mapping to colors. \code{"all"} applies colors to
span the entire range of the dataset (all rows, all colors). \code{"row"} or \code{"column"} normalizes
each row or column independently.}

\item{colorNA}{Color used for NAs in \code{data} or for reactions missing from \code{data}.}
}
\description{
Color a metabolic map.
}

